crawler/pyproject.toml
[project]
name = "job-crawler"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
]

[project.scripts]
job-crawler = "crawler.__main__:main"
